Use dyncallback to get real function pointers in interpreter
This is just on the desktop; for WASM, 'fakeExtern' casts an Operation* to a FunPtr and implements 'doDynCall' accordingly.

This makes it possible to use pthreads with the interpreter; it's no longer single-threaded. (Except in WASM which is always single-threaded.)
This gets demo/sound/main.crow working in the interpreter; previously this crashed when native code tried to invoke the interpreter's phony function pointer.

This has some side effects:

* It's no longer practical to use 'allFunPtrs' for backtrace. Instead implement backtracing using different strategies:
  - In interpreter: Use a builtin function 'interpreter-backtrace'
  - On Windows: Use 'DbgHelp.lib'
  - Otherwise: Use libunwind

* Since there may be multiple parallel interpreters, use 'static' thread-local variables. DMD does not work well with these on Windows (https://issues.dlang.org/show_bug.cgi?id=23024), use only LDC on Windows.